Refactor and remove cards from index.js
