+++ lib/index.js
@@ -1,24 +1,6 @@
 var $ = require('jquery')
-
-  var deck = [
-    { name: "A", suit: "Hearts", value: 11 },
-    { name: "K", suit: "Spades", value: 10 },
-    { name: "Q", suit: "Clubs", value: 10 },
-    { name: "Two", suit: "Hearts", value: 2 },
-    { name: "Four", suit: "Spades", value: 4 },
-    { name: "Seven", suit: "Clubs", value: 7 },
-    { name: "Jack", suit: "Hearts", value: 10 },
-    { name: "Three", suit: "Diamonds", value: 3 },
-    { name: "Ace", suit: "Diamonds", value: 11 },
-  ]
-
-  function pickCard () {
-    var random = Math.floor(Math.random()*deck.length)
-    var card = deck[random];
-    deck.splice(random, 1)
-    return card;
-  }
-
+var pickCard = require ('./deck')
+  
 function dealCards () {
   dealDealerHand($("#dealer-hand"));
   dealPlayerHand($("#player-hand"));
